+++ Shuhari.Framework.Common/Properties/SharedAssemblyInfo.cs
@@ -8,4 +8,4 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("0.1.10.1")]
+[assembly: AssemblyVersion("0.1.11.0")]